requirements.txt
boto3==1.14.28
dynamodb-json==1.3
PTable==0.9.2
python-telegram-bot==12.8
requests==2.24.0
